requirements.txt
Django==4.1.2
sqlparse==0.4.2
djangorestframework==3.14.0